Update Get-JLAPS.sh
was getting [0] first value of password array for expire time. Needed to use [-1] to get last item in array to make sure provided time is accurate for latest password. 
Alternative would be to perform a password match on the current password and go down a few lines from that, yet don't want to do that unless I need to.