Update GraphAPI_MapNetworkDrives_generalized.txt
Script intended to be used when Kerberos identity already exists on device. 
For instances where Mac devices are not joined to AD, yet there is AD sync setup with Entra and need to get group memberships for drive mappings.

Check kerberos tickets for SamAccount name and grab account and user information from Entra ID. As configured, wants accessToken as an argument to this script. Alternatively can configure with enterprise application with client and secret. Waiting on that, don't have the admin consent on needed permissions this was for testing in preparation for that.

This script will map drives based on found matched group membership(s)