Update uploadJamfLogs
Added function to check for existing attachments, log them, and delete them. 

This only runs if the purge variable is set (not empty). I could probably expand this logic to only run for specific file types. Yet not doing that today...
To not conflict with something like uploaded pdfd for proof of purchase or applecare information.